blog项目，全局搜索、分页优化、专题详情和文章详情

+++ src/main/java/com/scs/web/blog/domain/vo/UserVo.java
@@ -29,9 +29,9 @@ public class UserVo {
     private String banner;
     private String email;
     private String homepage;
-    private Short follows;
-    private Short fans;
-    private Short articles;
+    private Integer follows;
+    private Integer fans;
+    private Integer articles;
     private LocalDateTime createTime;
     private Short status;
     private List<Article> articleList;
